+++ src/creature/brain/neuron/internal.rs
@@ -28,7 +28,7 @@ impl InternalNeuron {
 
 impl Activation for InternalNeuron {
     // TODO: make implementation of Activation the same for ActionNeuron and InternalNeuron
-    fn activation(&self, internal_activation_cache: &mut HashMap<Arc<InternalNeuron>, f64>) -> f64 {
+    fn activation(&self, internal_activation_cache: &mut HashMap<Arc<InternalNeuron>, f32>) -> f32 {
         let activation = self
             .inputs()
             .iter()
@@ -50,7 +50,7 @@ impl Activation for InternalNeuron {
                     connection.weight() * sensory_neuron.activation(internal_activation_cache)
                 }
             })
-            .sum::<f64>()
+            .sum::<f32>()
             .tanh();
 
         return activation;
